feat(routes): full routes implementation
