+++ src/server.js
@@ -1,35 +1,22 @@
 // node recommends to still import;
 // https://nodejs.org/api/buffer.html#buffer
 import { createServer } from 'node:http';
-import { randomUUID } from 'node:crypto';
 import { json } from './middlewares/json.js';
-import { Database } from './db.js';
-
-const database = new Database();
+import { routes } from './routes.js';
 
 const server = createServer(async (req, res) => {
   const { method, url } = req;
-
   await json(req, res);
 
-  if (method === 'GET' && url === '/users') {
-    const users = database.select('users');
-    return res.end(JSON.stringify(users));
+  const route = routes.find((route) => {
+    return route.method === method && route.path === url;
+  });
+
+  if (route) {
+    return route.handler(req, res);
   }
 
-  if (method === 'POST' && url === '/users') {
-    const { name, email } = req.body;
-
-    const user = {
-      id: randomUUID(),
-      name,
-      email,
-    };
-    await database.insert('users', user);
-    return res.writeHead(201).end();
-  }
-
-  return res.writeHead(404).end('No routes found');
+  return res.writeHead(404).end();
 });
 
 server.listen(5000, () => {
